Struct lru_time_cache::LruCache

pub struct LruCache<Key, Value> { /* fields omitted */ }

Implementation of LRU cache.

Implementations

impl<Key, Value> LruCache<Key, Value> where
Key: Ord + Clone,

pub fn with_capacity(capacity: usize) -> LruCache<Key, Value>

Constructor for capacity based LruCache.

pub fn with_expiry_duration(time_to_live: Duration) -> LruCache<Key, Value>

Constructor for time based LruCache.

pub fn with_expiry_duration_and_capacity(
time_to_live: Duration,
capacity: usize
) -> LruCache<Key, Value>

Constructor for dual-feature capacity and time based LruCache.

pub fn notify_insert(
&mut self,
key: Key,
value: Value
) -> (Option<Value>, Vec<(Key, Value)>)

Inserts a key-value pair into the cache.

If the key already existed in the cache, the existing value is returned and overwritten in the cache. Otherwise, the key-value pair is inserted and None is returned. Evicts and returns expired entries.

pub fn insert(&mut self, key: Key, value: Value) -> Option<Value>

Inserts a key-value pair into the cache.

If the key already existed in the cache, the existing value is returned and overwritten in the cache. Otherwise, the key-value pair is inserted and None is returned.

pub fn remove<Q: ?Sized>(&mut self, key: &Q) -> Option<Value> where
Key: Borrow<Q>,
Q: Ord,

Removes a key-value pair from the cache.

pub fn clear(&mut self)

Clears the LruCache, removing all values.

pub fn notify_get<Q: ?Sized>(
&mut self,
key: &Q
) -> (Option<&Value>, Vec<(Key, Value)>) where
Key: Borrow<Q>,
Q: Ord,

Much like get(), except in addition returns expired entries.

pub fn get<Q: ?Sized>(&mut self, key: &Q) -> Option<&Value> where
Key: Borrow<Q>,
Q: Ord,

Retrieves a reference to the value stored under key, or None if the key doesn't exist. Also removes expired elements and updates the time.

pub fn peek<Q: ?Sized>(&self, key: &Q) -> Option<&Value> where
Key: Borrow<Q>,
Q: Ord,

Returns a reference to the value with the given key, if present and not expired, without updating the timestamp.

pub fn notify_get_mut<Q: ?Sized>(
&mut self,
key: &Q
) -> (Option<&mut Value>, Vec<(Key, Value)>) where
Key: Borrow<Q>,
Q: Ord,

Retrieves a mutable reference to the value stored under key, or None if the key doesn't exist. Also removes expired elements and updates the time.

pub fn get_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<&mut Value> where
Key: Borrow<Q>,
Q: Ord,

Retrieves a mutable reference to the value stored under key, or None if the key doesn't exist. Also removes expired elements and updates the time.

pub fn contains_key<Q: ?Sized>(&self, key: &Q) -> bool where
Key: Borrow<Q>,
Q: Ord,

Returns whether key exists in the cache or not.

pub fn len(&self) -> usize

Returns the size of the cache, i.e. the number of cached non-expired key-value pairs.

pub fn is_empty(&self) -> bool

Returns true if there are no non-expired entries in the cache.

pub fn entry(&mut self, key: Key) -> Entry<'_, Key, Value>

Gets the given key's corresponding entry in the map for in-place manipulation.

pub fn notify_iter(&mut self) -> NotifyIter<'_, Key, Value>

Returns an iterator over all entries that updates the timestamps as values are traversed. Also removes expired elements before creating the iterator. Values are produced in the most recently used order.

Also, evicts and returns expired entries.

pub fn iter(&mut self) -> Iter<'_, Key, Value>

Returns an iterator over all entries that updates the timestamps as values are traversed. Also removes expired elements before creating the iterator. Values are produced in the most recently used order.

pub fn peek_iter(&self) -> PeekIter<'_, Key, Value>

Returns an iterator over all entries that does not modify the timestamps.