Struct lru_time_cache::LruCache

pub struct LruCache<K, V> {
    // some fields omitted
}

Provides a Last Recently Used caching algorithm in a container which may be limited by size or time, reordered to most recently seen.

Methods

impl<K, V> LruCache<K, V> where K: PartialOrd + Ord + Clone, V: Clone

Constructor for size (capacity) based LruCache

fn with_capacity(capacity: usize) -> LruCache<K, V>

Constructor for size based LruCache

fn with_expiry_duration(time_to_live: Duration) -> LruCache<K, V>

Constructor for time based LruCache

fn with_expiry_duration_and_capacity(time_to_live: Duration, capacity: usize) -> LruCache<K, V>

Constructor for dual feature capacity, or time based LruCache

fn insert(&mut self, key: K, value: V) -> Option<V>

Inserts a key-value pair into the cache. If the key already had a value present in the cache, that value is returned. Otherwise, None is returned.

fn remove(&mut self, key: &K) -> Option<V>

Remove a key/value pair from cache

fn get(&mut self, key: &K) -> Option<&V>

Retrieve a value from cache

fn get_mut(&mut self, key: &K) -> Option<&mut V>

Retrieve a value from cache

fn contains_key(&mut self, key: &K) -> bool

Returns true if a value existed for the specified key.

fn len(&mut self) -> usize

Current size of cache

fn retrieve_all(&mut self) -> Vec<(K, V)>

Retrieve all elements as a vector of key value tuple.

fn retrieve_all_ordered(&mut self) -> Vec<(K, V)>

Return a vector of key value pairs ordered by most to least recently updated.

fn entry(&mut self, key: K) -> Entry<K, V>

Gets the given key's corresponding entry in the map for in-place manipulation.